+++ bika/lims/browser/calcs.py
@@ -22,7 +22,7 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.browser import BrowserView
 from bika.lims.interfaces import IAnalysis
 from bika.lims.interfaces import IFieldIcons
-from bika.lims.utils import convert_unit
+from bika.lims.utils import resolve_unit
 from bika.lims.utils import t, isnumber
 from bika.lims.utils.analysis import format_numeric_result
 
@@ -74,7 +74,7 @@ class ajaxCalculateAnalysisEntry(BrowserView):
         self.context = context
         self.request = request
 
-    def calculate(self, uid=None, sample_type_uid=None):
+    def calculate(self, uid=None):
         analysis = self.analyses[uid]
         form_result = self.current_results[uid]['result']
         calculation = analysis.getCalculation()
@@ -95,23 +95,8 @@ class ajaxCalculateAnalysisEntry(BrowserView):
             if form_result == "0/0":
                 Result['result'] = ""
 
-        if sample_type_uid and Result['result']:
-            service = analysis.getAnalysisService()
-            for unit_conversion in service.getUnitConversions():
-                if unit_conversion.get('SampleType') and \
-                   unit_conversion.get('Unit') and \
-                   unit_conversion.get('ShowOnListing', False) and \
-                   unit_conversion.get('SampleType') == sample_type_uid:
-                    conversion = ploneapi.content.get(
-                            UID=unit_conversion['Unit'])
-                    Result['converted_result'] = '%s %s' % (
-                        convert_unit(
-                            Result['result'],
-                            conversion.formula,
-                            self.context.bika_setup.getResultsDecimalMark(),
-                            analysis.getPrecision()),
-                        conversion.converted_unit)
-                    break
+        if Result['result']:
+            Result['converted_result'] = resolve_unit(analysis, Result['result'])
 
         if calculation:
             """We need first to create the map of available parameters
@@ -398,8 +383,7 @@ class ajaxCalculateAnalysisEntry(BrowserView):
             self.analyses[analysis_uid] = analysis
 
         if uid not in self.ignore_uids:
-            sample_type_uid = self.context.getSampleType().UID()
-            self.calculate(uid, sample_type_uid)
+            self.calculate(uid)
 
         results = []
         for result in self.results: