+++ bika/lims/content/abstractbaseanalysis.py
@@ -15,6 +15,11 @@ from Products.Archetypes.Widget import BooleanWidget, DecimalWidget, \
     IntegerWidget, SelectionWidget, StringWidget
 from Products.Archetypes.utils import DisplayList, IntDisplayList
 from Products.CMFCore.utils import getToolByName
+from Products.DataGridField import CheckboxColumn
+from Products.DataGridField import Column
+from Products.DataGridField import DataGridField
+from Products.DataGridField import DataGridWidget
+from Products.DataGridField import SelectColumn
 from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.fields import DurationField, UIDReferenceField
 from bika.lims.browser.widgets.durationwidget import DurationWidget
@@ -25,6 +30,7 @@ from bika.lims.config import ATTACHMENT_OPTIONS, SERVICE_POINT_OF_CAPTURE
 from bika.lims.content.bikaschema import BikaSchema
 from bika.lims.interfaces import IBaseAnalysis
 from bika.lims.utils import to_utf8 as _c
+from bika.lims.vocabularies import CatalogVocabulary
 from zope.interface import implements
 
 # Anywhere that there just isn't space for unpredictably long names,
@@ -653,6 +659,42 @@ Remarks = TextField(
     schemata='Description'
 )
 
+# This field stores unit conversions for each each sample type on 
+# the Result Options tab
+UnitConversions = DataGridField(
+    'UnitConversions',
+    schemata="Result Options",
+    allow_insert=True,
+    allow_delete=True,
+    allow_reorder=True,
+    allow_empty_rows=False,
+    columns=('SampleType',
+             'HidePrimaryUnit',
+             'ShowOnListing',
+             'Unit',),
+    default=[{'SampleType': [],
+              'Unit': '',
+              'HidePrimaryUnit': '0',
+              }],
+    widget=DataGridWidget(
+        label=_("Reporting Units per Sample Type"),
+        description=_("Analyses can be reported with more that one unit. Here is the list of additional reporting units per Sample Type."),
+        columns={
+            'SampleType': SelectColumn(
+                'Sample Type',
+                vocabulary='Vocabulary_SampleTypes'),
+            'HidePrimaryUnit': CheckboxColumn(
+                'Hide Primary Unit',),
+            'ShowOnListing': CheckboxColumn(
+                'Show On Listing',),
+            'Unit': SelectColumn(
+                'Unit',
+                vocabulary='Vocabulary_UnitConversions'),
+        }
+    )
+)
+
+
 schema = BikaSchema.copy() + Schema((
     ShortTitle,
     SortKey,
@@ -689,7 +731,8 @@ schema = BikaSchema.copy() + Schema((
     Hidden,
     SelfVerification,
     NumberOfRequiredVerifications,
-    Remarks
+    Remarks,
+    UnitConversions
 ))
 
 schema['id'].widget.visible = False
@@ -967,3 +1010,23 @@ class AbstractBaseAnalysis(BaseContent):  # TODO BaseContent?  is really needed?
         department = self.getDepartment()
         if department:
             return department.UID()
+
+    @security.public
+    def Vocabulary_SampleTypes(self):
+        vocabulary = CatalogVocabulary(self)
+        vocabulary.catalog = 'bika_setup_catalog'
+        return vocabulary(allow_blank=True, portal_type='SampleType')
+
+    @security.public
+    def Vocabulary_UnitConversions(self):
+        catalog = getToolByName(self, 'portal_catalog')
+        brains = catalog(portal_type='UnitConversion')
+        items = [('', '')]
+        for brain in brains:
+            obj = brain.getObject()
+            key = obj.UID()
+            value = '{} --> {}'.format(
+                _c(obj.title), _c(obj.converted_unit))
+            items.append((key, value))
+        return DisplayList(items)
+